requirements.txt
flask~=2.2.3
pyjwt~=2.6.0
requests~=2.28.2
